Added function to start CAN RX task
The stack size for the CAN task was hardcoded into the csp_can_init.
Instead we provide a function to start the task so the stack size
becomes an agument. Furthermore the recent changed that added two
distinct functions:
  csp_can_process_frame() and
  csp_can_process_frame_deferred()
enables the possibility of calling the process_frame() directly from ISR
and thereby avoid having to allocate a full task to do it.