Fix a waste of memory
On POSIX systems, a memory space (288 bytes for Linux) is reserved by each threads to return informations while they terminate. This memory space is not freed until pthread_join() is called on them and return. As CSP does not include any function to join a thread, this is a memory waste at the best, or a memory leak if threads are started and stopped as a regular operation. Furthermore, Valgrind detect this issue as a memory leak and prevent users to reach the "valgrind-clean" status.
Setting the thread as detached does not allocate this memory space and avoid any trouble.
At the same time, we can reuse the attribute structure to take in consideration the stack size provided by users.